Allow dragging floating AUI frames by the gripper.
Previously starting to drag a pane by its gripper worked but once it undocked
it was impossible to drag it by the gripper any more which was annoying for
the user and unexpected.

Fix this by ensuring that we move the correct top level window and not the
child window when dragging floating panes.

Closes #13177.

git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@67751 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775